+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Deprecations
 
 * Deprecated the `Command::question()` method. Use the `Command::input()` method instead.
+* Deprecated the `Cursor::beginningOfLine()` method. Use the `Cursor::moveToBeginningOfLine()` method instead.
 
 > All deprecated features will be removed in Mako 12.0.